+++ ffi/internal/ffi_macro.nim
@@ -2361,12 +2361,12 @@ when defined(ffiGenBindings):
     of "rust":
       generateRustCrate(
         genProcs, ffiTypeRegistry, libName, outDir, srcRel, ffiEventRegistry,
-        ffiConstRegistry,
+        ffiConstRegistry, ffiReverseRegistry, ffiReverseEventRegistry,
       )
     of "cpp", "c++":
       generateCppBindings(
         genProcs, ffiTypeRegistry, libName, outDir, srcRel, ffiEventRegistry,
-        ffiConstRegistry,
+        ffiConstRegistry, ffiReverseRegistry, ffiReverseEventRegistry,
       )
     of "c":
       generateCBindings(